fix migrations. start work on code refactor
